requirements.txt
langchain_openai
langchain_core
langchain_community
streamlit
python-dotenv
langserve
fastapi
uvicorn
sse_starlette